Isolate document updating logic
