+++ apps/meteor/client/lib/cachedCollections/LocalCollection.ts
@@ -1,7 +1,7 @@
-import { createComparatorFromSort, createPredicateFromFilter } from '@rocket.chat/mongo-adapter';
+import { createPredicateFromFilter } from '@rocket.chat/mongo-adapter';
 import type { FieldExpression, Filter } from '@rocket.chat/mongo-adapter';
 import { Meteor } from 'meteor/meteor';
-import type { CountDocumentsOptions, UpdateFilter } from 'mongodb';
+import type { CountDocumentsOptions } from 'mongodb';
 import type { StoreApi, UseBoundStore } from 'zustand';
 
 import { Cursor } from './Cursor';
@@ -12,18 +12,16 @@ import { Matcher } from './Matcher';
 import { MinimongoError } from './MinimongoError';
 import type { Query } from './Query';
 import { SynchronousQueue } from './SynchronousQueue';
-import {
-	hasOwn,
-	isIndexable,
-	isNumericKey,
-	isOperatorObject,
-	populateDocumentWithQueryFields,
-	_f,
-	_isPlainObject,
-	_selectorIsId,
-	clone,
-} from './common';
+import type { UpdateFilter } from './Updater';
+import { Updater } from './Updater';
+import type { ArrayIndices } from './common';
+import { hasOwn, _selectorIsId, clone, assertHasValidFieldNames } from './common';
 
+/**
+ * Forked from Meteor's Mongo.Collection, this class implements a local collection over a Zustand store.
+ *
+ * Do not use this class directly.
+ */
 export class LocalCollection<T extends { _id: string }> {
 	readonly observeQueue = new SynchronousQueue();
 
@@ -670,11 +668,12 @@ export class LocalCollection<T extends { _id: string }> {
 		return matchedBefore;
 	}
 
-	private _modifyAndNotify(doc: T, mod: UpdateFilter<T>, arrayIndices: (number | 'x')[] | undefined) {
+	private _modifyAndNotify(doc: T, mod: UpdateFilter<T>, arrayIndices: ArrayIndices | undefined) {
 		const matchedBefore = this._getMatchedDocAndModify(doc);
 
 		const oldDoc = clone(doc);
-		doc = this._modify(doc, mod, { arrayIndices });
+		const updater = new Updater(clone(mod));
+		doc = updater.modify(doc, { arrayIndices });
 		this.store.setState((state) => ({
 			records: state.records.map((record) => (record._id === doc._id ? doc : record)),
 		}));
@@ -702,11 +701,12 @@ export class LocalCollection<T extends { _id: string }> {
 		return recomputeQueries;
 	}
 
-	private async _modifyAndNotifyAsync(doc: T, mod: UpdateFilter<T>, arrayIndices: (number | 'x')[] | undefined) {
+	private async _modifyAndNotifyAsync(doc: T, mod: UpdateFilter<T>, arrayIndices: ArrayIndices | undefined) {
 		const matchedBefore = this._getMatchedDocAndModify(doc);
 
 		const oldDoc = clone(doc);
-		doc = this._modify(doc, mod, { arrayIndices });
+		const updater = new Updater(clone(mod));
+		doc = updater.modify(doc, { arrayIndices });
 		this.store.setState((state) => ({
 			records: state.records.map((record) => (record._id === doc._id ? doc : record)),
 		}));
@@ -786,30 +786,9 @@ export class LocalCollection<T extends { _id: string }> {
 		return first;
 	}
 
-	private _createUpsertDocument(selector: Filter<T>, modifier: UpdateFilter<T>) {
-		const selectorDocument = populateDocumentWithQueryFields(selector);
-		const isModify = this._isModificationMod(modifier);
-
-		let newDoc: Partial<T> = {};
-
-		if (selectorDocument._id) {
-			newDoc._id = selectorDocument._id;
-			delete selectorDocument._id;
-		}
-
-		newDoc = this._modify(newDoc, { $set: selectorDocument });
-		newDoc = this._modify(newDoc, modifier, { isInsert: true });
-
-		if (isModify) {
-			return newDoc as T;
-		}
-
-		const replacement = Object.assign({}, modifier);
-		if (newDoc._id) {
-			replacement._id = newDoc._id;
-		}
-
-		return replacement as T;
+	private _createUpsertDocument(selector: Filter<T>, modifier: UpdateFilter<T>): T {
+		const updater = new Updater(modifier);
+		return updater.createUpsertDocument(selector);
 	}
 
 	private _findInOrderedResults(query: Query<T>, doc: T): number {
@@ -876,89 +855,6 @@ export class LocalCollection<T extends { _id: string }> {
 		array.splice(i, 0, value);
 
 		return i;
-	}
-
-	private _isModificationMod(mod: UpdateFilter<T> | T) {
-		let isModify = false;
-		let isReplace = false;
-
-		for (const key of Object.keys(mod)) {
-			if (key.slice(0, 1) === '$') {
-				isModify = true;
-			} else {
-				isReplace = true;
-			}
-		}
-
-		if (isModify && isReplace) {
-			throw new MinimongoError('Update parameter cannot have both modifier and non-modifier fields.');
-		}
-
-		return isModify;
-	}
-
-	private _modify<U extends Partial<T>>(
-		doc: U,
-		modifier: UpdateFilter<T>,
-		options: { isInsert?: boolean; arrayIndices?: (number | 'x')[] } = {},
-	) {
-		if (!_isPlainObject(modifier)) {
-			throw new MinimongoError('Modifier must be an object');
-		}
-
-		modifier = clone(modifier);
-
-		const isModifier = isOperatorObject(modifier);
-
-		if (isModifier) {
-			const newDoc = clone(doc);
-
-			for (const [operator, operand] of Object.entries(modifier)) {
-				const modFunc = MODIFIERS[options.isInsert && operator === '$setOnInsert' ? '$set' : (operator as keyof typeof MODIFIERS)];
-
-				if (!modFunc) {
-					throw new MinimongoError(`Invalid modifier specified ${operator}`);
-				}
-
-				for (const [keypath, arg] of Object.entries(operand)) {
-					if (keypath === '') {
-						throw new MinimongoError('An empty update path is not valid.');
-					}
-
-					const keyparts = keypath.split('.');
-
-					if (!keyparts.every(Boolean)) {
-						throw new MinimongoError(`The update path '${keypath}' contains an empty field name, which is not allowed.`);
-					}
-
-					const target = findModTarget(newDoc, keyparts, {
-						arrayIndices: options.arrayIndices,
-						forbidArray: operator === '$rename',
-						noCreate: NO_CREATE_MODIFIERS[operator as keyof typeof NO_CREATE_MODIFIERS],
-					});
-
-					modFunc(target as any, keyparts.pop()!, arg, keypath, newDoc);
-				}
-			}
-
-			if (doc._id && doc._id !== newDoc._id) {
-				throw new MinimongoError(
-					`After applying the update to the document {_id: "${doc._id}", ...},` +
-						" the (immutable) field '_id' was found to have been altered to " +
-						`_id: "${newDoc._id}"`,
-				);
-			}
-
-			return newDoc;
-		}
-
-		if (doc._id && modifier._id && doc._id !== modifier._id) {
-			throw new MinimongoError(`The _id field cannot be changed from {_id: "${doc._id}"} to {_id: "${modifier._id}"}`);
-		}
-
-		assertHasValidFieldNames(modifier);
-
-		return modifier as T;
 	}
 
 	private _removeFromResults(query: Query<T>, doc: T) {
@@ -1064,477 +960,4 @@ export class LocalCollection<T extends { _id: string }> {
 			if (query.movedBefore) await query.movedBefore(doc._id, next);
 		}
 	}
-}
-
-const MODIFIERS = {
-	$currentDate(target: Record<string, Date>, field: string, arg: any) {
-		if (typeof arg === 'object' && hasOwn.call(arg, '$type')) {
-			if (arg.$type !== 'date') {
-				throw new MinimongoError('Minimongo does currently only support the date type in $currentDate modifiers', { field });
-			}
-		} else if (arg !== true) {
-			throw new MinimongoError('Invalid $currentDate modifier', { field });
-		}
-
-		target[field] = new Date();
-	},
-	$inc(target: Record<string, number>, field: string, arg: any) {
-		if (typeof arg !== 'number') {
-			throw new MinimongoError('Modifier $inc allowed for numbers only', { field });
-		}
-
-		if (field in target) {
-			if (typeof target[field] !== 'number') {
-				throw new MinimongoError('Cannot apply $inc modifier to non-number', { field });
-			}
-
-			target[field] += arg;
-		} else {
-			target[field] = arg;
-		}
-	},
-	$min(target: Record<string, number>, field: string, arg: any) {
-		if (typeof arg !== 'number') {
-			throw new MinimongoError('Modifier $min allowed for numbers only', { field });
-		}
-
-		if (field in target) {
-			if (typeof target[field] !== 'number') {
-				throw new MinimongoError('Cannot apply $min modifier to non-number', { field });
-			}
-
-			if (target[field] > arg) {
-				target[field] = arg;
-			}
-		} else {
-			target[field] = arg;
-		}
-	},
-	$max(target: Record<string, number>, field: string, arg: any) {
-		if (typeof arg !== 'number') {
-			throw new MinimongoError('Modifier $max allowed for numbers only', { field });
-		}
-
-		if (field in target) {
-			if (typeof target[field] !== 'number') {
-				throw new MinimongoError('Cannot apply $max modifier to non-number', { field });
-			}
-
-			if (target[field] < arg) {
-				target[field] = arg;
-			}
-		} else {
-			target[field] = arg;
-		}
-	},
-	$mul(target: Record<string, number>, field: string, arg: any) {
-		if (typeof arg !== 'number') {
-			throw new MinimongoError('Modifier $mul allowed for numbers only', { field });
-		}
-
-		if (field in target) {
-			if (typeof target[field] !== 'number') {
-				throw new MinimongoError('Cannot apply $mul modifier to non-number', { field });
-			}
-
-			target[field] *= arg;
-		} else {
-			target[field] = 0;
-		}
-	},
-	$rename(target: Record<string, object>, field: string, arg: any, keypath: any, doc: any) {
-		if (keypath === arg) {
-			throw new MinimongoError('$rename source must differ from target', { field });
-		}
-
-		if (target === null) {
-			throw new MinimongoError('$rename source field invalid', { field });
-		}
-
-		if (typeof arg !== 'string') {
-			throw new MinimongoError('$rename target must be a string', { field });
-		}
-
-		if (arg.includes('\0')) {
-			throw new MinimongoError("The 'to' field for $rename cannot contain an embedded null byte", { field });
-		}
-
-		if (target === undefined) {
-			return;
-		}
-
-		const object = target[field];
-
-		delete target[field];
-
-		const keyparts = arg.split('.');
-		const target2 = findModTarget(doc, keyparts, { forbidArray: true });
-
-		if (!target2) {
-			throw new MinimongoError('$rename target field invalid', { field });
-		}
-
-		target2[keyparts.pop() as keyof typeof target2] = object;
-	},
-	$set(target: Record<string, unknown>, field: string, arg: any) {
-		if (target !== Object(target)) {
-			const error = new MinimongoError('Cannot set property on non-object field', { field });
-			error.setPropertyError = true;
-			throw error;
-		}
-
-		if (target === null) {
-			const error = new MinimongoError('Cannot set property on null', { field });
-			error.setPropertyError = true;
-			throw error;
-		}
-
-		assertHasValidFieldNames(arg);
-
-		target[field] = arg;
-	},
-	$setOnInsert() {
-		throw new MinimongoError('It should have been converted to $set in _modify');
-	},
-	$unset(target: Record<string, unknown>, field: string) {
-		if (target !== undefined) {
-			if (Array.isArray(target)) {
-				if (field in target) {
-					target[field] = null;
-				}
-			} else {
-				delete target[field];
-			}
-		}
-	},
-	$push(target: Record<string, { _id: string }[]>, field: string, arg: any) {
-		if (target[field] === undefined) {
-			target[field] = [];
-		}
-
-		if (!Array.isArray(target[field])) {
-			throw new MinimongoError('Cannot apply $push modifier to non-array', { field });
-		}
-
-		if (!arg?.$each) {
-			assertHasValidFieldNames(arg);
-
-			target[field].push(arg);
-
-			return;
-		}
-
-		const toPush = arg.$each;
-		if (!Array.isArray(toPush)) {
-			throw new MinimongoError('$each must be an array', { field });
-		}
-
-		assertHasValidFieldNames(toPush);
-
-		let position = undefined;
-		if ('$position' in arg) {
-			if (typeof arg.$position !== 'number') {
-				throw new MinimongoError('$position must be a numeric value', { field });
-			}
-
-			if (arg.$position < 0) {
-				throw new MinimongoError('$position in $push must be zero or positive', { field });
-			}
-
-			position = arg.$position;
-		}
-
-		let slice = undefined;
-		if ('$slice' in arg) {
-			if (typeof arg.$slice !== 'number') {
-				throw new MinimongoError('$slice must be a numeric value', { field });
-			}
-
-			slice = arg.$slice;
-		}
-
-		let sortFunction = undefined;
-		if (arg.$sort) {
-			if (slice === undefined) {
-				throw new MinimongoError('$sort requires $slice to be present', { field });
-			}
-
-			sortFunction = createComparatorFromSort(arg.$sort);
-
-			for (const element of toPush) {
-				if (_f._type(element) !== 3) {
-					throw new MinimongoError('$push like modifiers using $sort require all elements to be objects', { field });
-				}
-			}
-		}
-
-		if (position === undefined) {
-			for (const element of toPush) {
-				target[field].push(element);
-			}
-		} else {
-			const spliceArguments = [position, 0];
-
-			for (const element of toPush) {
-				spliceArguments.push(element);
-			}
-
-			target[field].splice(...(spliceArguments as Parameters<typeof Array.prototype.splice>));
-		}
-
-		if (sortFunction) {
-			target[field].sort(sortFunction);
-		}
-
-		if (slice !== undefined) {
-			if (slice === 0) {
-				target[field] = [];
-			} else if (slice < 0) {
-				target[field] = target[field].slice(slice);
-			} else {
-				target[field] = target[field].slice(0, slice);
-			}
-		}
-	},
-	$pushAll(target: Record<string, unknown[]>, field: string, arg: any) {
-		if (!(typeof arg === 'object' && Array.isArray(arg))) {
-			throw new MinimongoError('Modifier $pushAll/pullAll allowed for arrays only');
-		}
-
-		assertHasValidFieldNames(arg);
-
-		const toPush = target[field];
-
-		if (toPush === undefined) {
-			target[field] = arg;
-		} else if (!Array.isArray(toPush)) {
-			throw new MinimongoError('Cannot apply $pushAll modifier to non-array', { field });
-		} else {
-			toPush.push(...arg);
-		}
-	},
-	$addToSet(target: Record<string, unknown[]>, field: string, arg: any) {
-		let isEach = false;
-
-		if (typeof arg === 'object') {
-			const keys = Object.keys(arg);
-			if (keys[0] === '$each') {
-				isEach = true;
-			}
-		}
-
-		const values = isEach ? arg.$each : [arg];
-
-		assertHasValidFieldNames(values);
-
-		const toAdd = target[field];
-		if (toAdd === undefined) {
-			target[field] = values;
-		} else if (!Array.isArray(toAdd)) {
-			throw new MinimongoError('Cannot apply $addToSet modifier to non-array', { field });
-		} else {
-			for (const value of values) {
-				if (toAdd.some((element) => _f._equal(value, element))) {
-					continue;
-				}
-
-				toAdd.push(value);
-			}
-		}
-	},
-	$pop(target: Record<string, unknown[]>, field: string, arg: any) {
-		if (target === undefined) {
-			return;
-		}
-
-		const toPop = target[field];
-
-		if (toPop === undefined) {
-			return;
-		}
-
-		if (!Array.isArray(toPop)) {
-			throw new MinimongoError('Cannot apply $pop modifier to non-array', { field });
-		}
-
-		if (typeof arg === 'number' && arg < 0) {
-			toPop.splice(0, 1);
-		} else {
-			toPop.pop();
-		}
-	},
-	$pull(target: Record<string, { _id: string }[]>, field: string, arg: any) {
-		if (target === undefined) {
-			return;
-		}
-
-		const toPull = target[field];
-		if (toPull === undefined) {
-			return;
-		}
-
-		if (!Array.isArray(toPull)) {
-			throw new MinimongoError('Cannot apply $pull/pullAll modifier to non-array', { field });
-		}
-
-		let out;
-		if (arg != null && typeof arg === 'object' && !Array.isArray(arg)) {
-			const predicate = createPredicateFromFilter(arg);
-
-			out = toPull.filter((element) => !predicate(element));
-		} else {
-			out = toPull.filter((element) => !_f._equal(element, arg));
-		}
-
-		target[field] = out;
-	},
-	$pullAll(target: Record<string, unknown[]>, field: string, arg: any) {
-		if (!(typeof arg === 'object' && Array.isArray(arg))) {
-			throw new MinimongoError('Modifier $pushAll/pullAll allowed for arrays only', { field });
-		}
-
-		if (target === undefined) {
-			return;
-		}
-
-		const toPull = target[field];
-
-		if (toPull === undefined) {
-			return;
-		}
-
-		if (!Array.isArray(toPull)) {
-			throw new MinimongoError('Cannot apply $pull/pullAll modifier to non-array', { field });
-		}
-
-		target[field] = toPull.filter((object) => !arg.some((element) => _f._equal(object, element)));
-	},
-	$bit(_target: unknown, field: string) {
-		throw new MinimongoError('$bit is not supported', { field });
-	},
-	$v(_target: unknown, field: string) {
-		throw new MinimongoError('$v is not supported', { field });
-	},
-};
-
-const NO_CREATE_MODIFIERS = {
-	$pop: true,
-	$pull: true,
-	$pullAll: true,
-	$rename: true,
-	$unset: true,
-};
-
-const invalidCharMsg = {
-	'$': "start with '$'",
-	'.': "contain '.'",
-	'\0': 'contain null bytes',
-};
-
-function assertHasValidFieldNames(doc: unknown) {
-	if (doc && typeof doc === 'object') {
-		JSON.stringify(doc, (key, value) => {
-			assertIsValidFieldName(key);
-			return value;
-		});
-	}
-}
-
-function assertIsValidFieldName(key: string) {
-	let match;
-	if (typeof key === 'string' && (match = key.match(/^\$|\.|\0/))) {
-		throw new MinimongoError(`Key ${key} must not ${invalidCharMsg[match[0] as keyof typeof invalidCharMsg]}`);
-	}
-}
-
-function findModTarget(
-	doc: Record<string, any> | unknown[],
-	keyparts: (number | string)[],
-	options: {
-		noCreate?: boolean;
-		forbidArray?: boolean;
-		arrayIndices?: (number | 'x')[];
-	} = {},
-) {
-	let usedArrayIndex = false;
-
-	for (let i = 0; i < keyparts.length; i++) {
-		const last = i === keyparts.length - 1;
-		let keypart: string | number = keyparts[i];
-
-		if (!isIndexable(doc)) {
-			if (options.noCreate) {
-				return undefined;
-			}
-
-			const error = new MinimongoError(`cannot use the part '${keypart}' to traverse ${doc}`);
-			error.setPropertyError = true;
-			throw error;
-		}
-
-		if (Array.isArray(doc)) {
-			if (options.forbidArray) {
-				return null;
-			}
-
-			if (keypart === '$') {
-				if (usedArrayIndex) {
-					throw new MinimongoError("Too many positional (i.e. '$') elements");
-				}
-
-				if (!options.arrayIndices?.length) {
-					throw new MinimongoError('The positional operator did not find the match needed from the query');
-				}
-
-				keypart = options.arrayIndices[0];
-				usedArrayIndex = true;
-			} else if (isNumericKey(keypart as string)) {
-				keypart = parseInt(keypart as string);
-			} else {
-				if (options.noCreate) {
-					return undefined;
-				}
-
-				throw new MinimongoError(`can't append to array using string field name [${keypart}]`);
-			}
-
-			if (last) {
-				keyparts[i] = keypart;
-			}
-
-			if (options.noCreate && (keypart as number) >= doc.length) {
-				return undefined;
-			}
-
-			while (doc.length < (keypart as number)) {
-				doc.push(null);
-			}
-
-			if (!last) {
-				if (doc.length === keypart) {
-					doc.push({});
-				} else if (typeof doc[keypart as number] !== 'object') {
-					throw new MinimongoError(`can't modify field '${keyparts[i + 1]}' of list value ${JSON.stringify(doc[keypart as number])}`);
-				}
-			}
-		} else {
-			assertIsValidFieldName(keypart as string);
-
-			if (!(keypart in doc)) {
-				if (options.noCreate) {
-					return undefined;
-				}
-
-				if (!last) {
-					doc[keypart] = {};
-				}
-			}
-		}
-
-		if (last) return doc;
-
-		doc = doc[keypart as keyof typeof doc];
-	}
-
-	throw new MinimongoError('Should not reach here');
 }
